Game interface added
